Created form slice
