+++ src/inmanta/server/__init__.py
@@ -24,3 +24,4 @@ SLICE_SESSION_MANAGER = "core.session"
 SLICE_DATABASE = "core.database"
 SLICE_TRANSPORT = "core.transport"
 SLICE_COMPILER = "core.compiler"
+SLICE_FORM = "core.forms"
